DOUBLE

+++ stencil.c
@@ -130,8 +130,8 @@ void stencil(const int nx, const int ny,  float *restrict image, float *restrict
 
     // MPI_Sendrecv(lastRowSend, nx, MPI_FLOAT, rank + 1, 0, lastRowRecv, nx, MPI_FLOAT, rank +1, 0, MPI_COMM_WORLD, status);
 
-    MPI_Send(lastRowSend, nx, MPI_FLOAT,rank +1, 0, MPI_COMM_WORLD);
-    MPI_Recv(lastRowRecv, nx+1, MPI_FLOAT, rank+1, MASTER, MPI_COMM_WORLD, status);
+    MPI_Send(lastRowSend, nx, MPI_DOUBLE,rank +1, 0, MPI_COMM_WORLD);
+    MPI_Recv(lastRowRecv, nx+1, MPI_DOUBLE, rank+1, MASTER, MPI_COMM_WORLD, status);
 
     //Corner cases cmonnnnn
     tmp_image[0] = image[0] * 0.6f + (image[nx] + image[1]) * 0.1f; //comment
@@ -192,8 +192,8 @@ void stencil(const int nx, const int ny,  float *restrict image, float *restrict
 
     // MPI_Sendrecv(firstRowSend, nx, MPI_FLOAT, rank - 1, 0, firstRowRecv, nx, MPI_FLOAT, rank -1, 0, MPI_COMM_WORLD, status);
 
-    MPI_Recv(firstRowRecv, nx+1, MPI_FLOAT, rank-1 , MASTER, MPI_COMM_WORLD, status);
-    MPI_Send(firstRowSend, nx, MPI_FLOAT,rank -1, 0, MPI_COMM_WORLD);
+    MPI_Recv(firstRowRecv, nx+1, MPI_DOUBLE, rank-1 , MASTER, MPI_COMM_WORLD, status);
+    MPI_Send(firstRowSend, nx, MPI_DOUBLE,rank -1, 0, MPI_COMM_WORLD);
 
     //Corner cases cmonnnnn
     tmp_image[0] = image[0] * 0.6f + (image[nx] + image[1] + firstRowRecv[0]) * 0.1f; //comment
@@ -257,17 +257,17 @@ void stencil(const int nx, const int ny,  float *restrict image, float *restrict
 
     if (rank % 2 == 0)
     {
-      MPI_Send(firstRowSend, nx, MPI_FLOAT, rank - 1, 0, MPI_COMM_WORLD);
-      MPI_Send(lastRowSend, nx, MPI_FLOAT, rank + 1, 0, MPI_COMM_WORLD);
-      MPI_Recv(firstRowRecv, nx + 1, MPI_FLOAT, rank - 1, MASTER, MPI_COMM_WORLD, status);
-      MPI_Recv(lastRowRecv, nx + 1, MPI_FLOAT, rank + 1, MASTER, MPI_COMM_WORLD, status);
+      MPI_Send(firstRowSend, nx, MPI_DOUBLE, rank - 1, 0, MPI_COMM_WORLD);
+      MPI_Send(lastRowSend, nx, MPI_DOUBLE, rank + 1, 0, MPI_COMM_WORLD);
+      MPI_Recv(firstRowRecv, nx + 1, MPI_DOUBLE, rank - 1, MASTER, MPI_COMM_WORLD, status);
+      MPI_Recv(lastRowRecv, nx + 1, MPI_DOUBLE, rank + 1, MASTER, MPI_COMM_WORLD, status);
     }
     else
     {
-      MPI_Recv(firstRowRecv, nx + 1, MPI_FLOAT, rank - 1, MASTER, MPI_COMM_WORLD, status);
-      MPI_Recv(lastRowRecv, nx + 1, MPI_FLOAT, rank + 1, MASTER, MPI_COMM_WORLD, status);
-      MPI_Send(lastRowSend, nx, MPI_FLOAT, rank + 1, 0, MPI_COMM_WORLD);
-      MPI_Send(firstRowSend, nx, MPI_FLOAT, rank - 1, 0, MPI_COMM_WORLD);
+      MPI_Recv(firstRowRecv, nx + 1, MPI_DOUBLE, rank - 1, MASTER, MPI_COMM_WORLD, status);
+      MPI_Recv(lastRowRecv, nx + 1, MPI_DOUBLE, rank + 1, MASTER, MPI_COMM_WORLD, status);
+      MPI_Send(lastRowSend, nx, MPI_DOUBLE, rank + 1, 0, MPI_COMM_WORLD);
+      MPI_Send(firstRowSend, nx, MPI_DOUBLE, rank - 1, 0, MPI_COMM_WORLD);
     }
 
     // MPI_Sendrecv(firstRowSend, nx, MPI_FLOAT, rank - 1, 0, firstRowRecv, nx, MPI_FLOAT, rank -1, 0, MPI_COMM_WORLD, status);
